Add files via upload
Netflix Viewing Data
This dataset contains a record of my personal Netflix viewing history. The data includes information on the shows and movies I've watched, along with their genres, directors, release years, ratings, and other relevant attributes. This dataset is used for the Netflix-Personalised-Dashboard project to analyse viewing patterns, genre preferences, and content ratings.
Column Descriptions:
Title: The title of the show or movie.
Genre: The primary genre of the content (e.g., Drama, Action).
Director: The director(s) of the show or movie.
Country: The country where the content was produced.
Release Year: The year the show or movie was released.
Content Type: Indicates whether the content is a Movie or TV Show.
Duration: The length of the movie (in minutes) or the number of episodes for a TV show.
Rating: The age rating or content classification (e.g., PG-13, 18+).
IMDb Rating: The IMDb rating for the show or movie, if available.
Language: The primary language of the show or movie.
Date Watched: The date on which the content was viewed.